Expose transport pdu parser, split backend functions to own header
